Finishes tests for single fasta and adpats old ones so that they account for the fixed filtering but
