+++ backend/tests/protzilla/data_analysis/ptm_visualization/test_ptm_visualization.py
@@ -1,37 +1,31 @@
-import shutil
-from collections import Counter
-from contextlib import contextmanager
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Optional, Any
-from unittest import mock
 
 import pandas as pd
 import pytest
-from _pytest.monkeypatch import MonkeyPatch
 
-import main
-from protzilla.constants.intensity_types import IntensityType
 from protzilla.constants.paths import EXAMPLE_DATASET_METADATA_FILE
 from protzilla.data_analysis.ptm_visualization import (
     create_overview_ptm_visualization,
     create_bar_ptm_visualization,
     create_details_ptm_visualization,
-    ptm_vis_utils,
 )
 from protzilla.data_analysis.ptm_visualization.ptm_overview_plot import (
     get_detected_modifications,
 )
-from protzilla.data_analysis.ptm_visualization.ptm_vis_utils import (
-    get_general_config_module,
-)
-from protzilla.importing import peptide_import
-from protzilla.importing.metadata_import import metadata_import_method
 from tests.paths import (
     TEST_PTM_VISUALIZATION_PATH,
     TEST_FASTA_PATH,
     TEST_PEPTIDES_PATH,
     TEST_METADATA_PATH,
+)
+from tests.protzilla.data_analysis.ptm_visualization.ptm_vis_test_utils import (
+    get_evidence_df,
+    get_metadata_df,
+    alter_ptm_settings,
+    validate_plot_outputs,
+    mock_settings_file,
+    run_plot_and_validate,
 )
 
 GFAP_PATH = TEST_PTM_VISUALIZATION_PATH / "P14136"
@@ -59,34 +53,6 @@ def tmp_ptm_settings_dir(tmp_path_factory):
     test_tmp_data_dir = Path("ptm_settings/")
     tmp_path = tmp_path_factory.mktemp(str(test_tmp_data_dir))
     return tmp_path
-
-
-def get_evidence_df(path: Path):
-    outputs = peptide_import.evidence_import(
-        file_path=path,
-        intensity_name=IntensityType.INTENSITY.value,
-        map_to_uniprot=False,
-    )
-    evidence_df = outputs["peptide_df"]
-    return evidence_df
-
-
-def get_metadata_df(path: Path):
-    metadata_df = metadata_import_method(
-        protein_df=None, file_path=path, feature_orientation="columns"
-    )["metadata_df"]
-    return metadata_df
-
-
-def alter_ptm_settings(monkeypatch: MonkeyPatch, new_param_dict: dict):
-    def mock_figure_orientation(regions_file_path: Path, out_dir: Path):
-        config_module = get_general_config_module(regions_file_path, out_dir)
-        config_module.__dict__.update(new_param_dict)
-        return config_module
-
-    monkeypatch.setattr(
-        ptm_vis_utils, "get_general_config_module", mock_figure_orientation
-    )
 
 
 def pytest_generate_tests(metafunc):
@@ -163,148 +129,6 @@ class PlotValidationConfig:
     excluded_strings: tuple[str, ...] = ()
 
 
-def get_region_range_for_exon_coords(
-    exon_coords: list[tuple[float, float]], horizontal_orientation: bool
-) -> list[tuple[float, float]]:
-    # A bit of complicated logic to define the region range for the exon. The side of the exon that is longer, includes
-    # some buffer which is not actually part of the exon. Thus, we have to check which side is shorter and only include
-    # this one in the valid region ranges.
-
-    if not horizontal_orientation:
-        exon_coords = [(e[1], e[0]) for e in exon_coords]
-
-    new_regions = []
-    for x, y in exon_coords:
-        # figure out if the vertical line of the polygon is on the left or right side of the exon
-        # TODO: rename
-        blah = []
-        for i in range(len(y) - 1):
-            if y[i] != y[i + 1]:
-                blah.append(i)
-
-        if x[blah[0]] != x[blah[0] + 1]:
-            # vertical line at the beginning
-            new_regions.append(tuple(sorted(set(x))[:2]))
-        elif x[blah[1]] != x[blah[1] + 1]:
-            # vertical line at the end
-            new_regions.append(tuple(sorted(set(x))[1:]))
-        else:
-            assert False
-    return new_regions
-
-
-def validate_ptm_labels_in_bounds(plot):
-    # Does an alignment check of the PTMs to assert that they are not overflowing the sequence or are plotted in
-    # the exon gaps.
-    # We only do this for the overview plot, because in other plots we have way more shapes.
-    all_shapes = plot.layout.shapes
-    y_coords = {
-        tuple(sorted((shape.y0, shape.y1)))
-        for shape in all_shapes
-        if shape.type == "rect"
-    }
-    x_coords = {
-        tuple(sorted((shape.x0, shape.x1)))
-        for shape in all_shapes
-        if shape.type == "rect"
-    }
-    exon_coords = [
-        (el.x, el.y)
-        for el in plot.data
-        if el.text is None and len(el.x) == 5 and len(el.y) == 5
-    ]
-    if len(y_coords) == 1:
-        horizontal_orientation = True
-        valid_region_ranges = x_coords
-        valid_region_ranges.update(
-            get_region_range_for_exon_coords(exon_coords, horizontal_orientation)
-        )
-    elif len(x_coords) == 1:
-        horizontal_orientation = False
-        valid_region_ranges = y_coords
-        valid_region_ranges.update(
-            get_region_range_for_exon_coords(exon_coords, horizontal_orientation)
-        )
-    else:
-        raise AssertionError(
-            "Rects of the sequence have to be aligned in one of the dimensions"
-        )
-
-    label_lines = [
-        el
-        for el in plot.data
-        if el.text is None and len(el.x) == 2 and len(el.y) == 2
-    ]
-    for line in label_lines:
-        if horizontal_orientation:
-            assert len(set(line.x)) == 1
-            assert any(r[0] <= line.x[0] <= r[1] for r in valid_region_ranges)
-        else:
-            assert len(set(line.y)) == 1
-            assert any(r[0] <= line.y[0] <= r[1] for r in valid_region_ranges)
-
-
-def validate_plot_outputs(
-    plot,
-    plot_func,
-    all_groups: set,
-    required_groups: set,
-    validation_config: Optional[PlotValidationConfig],
-):
-    if plot_func == create_overview_ptm_visualization:
-        validate_ptm_labels_in_bounds(plot)
-
-    all_layout_strings = {anno.text for anno in plot.layout.annotations if anno.text}
-    all_data_strings = {
-        subplot.text
-        for subplot in plot.data
-        if hasattr(subplot, "mode") and subplot.mode == "text"
-    }
-    all_plot_strings = all_layout_strings.union(all_data_strings)
-
-    assert set(validation_config.required_ptm_types).issubset(all_plot_strings)
-    assert set(validation_config.required_ptms).issubset(all_plot_strings)
-    assert set(validation_config.required_region_names).issubset(all_plot_strings)
-
-    if plot_func in (create_details_ptm_visualization, create_bar_ptm_visualization):
-        required_groups = set(required_groups)
-        assert required_groups.issubset(all_plot_strings)
-        excluded_groups = all_groups - required_groups
-        assert all(g not in all_plot_strings for g in excluded_groups)
-
-    if plot_func == create_details_ptm_visualization:
-        assert set(validation_config.required_cleavages).issubset(all_plot_strings)
-        assert set(validation_config.required_region_short_names).issubset(
-            all_plot_strings
-        )
-
-    assert all(s not in all_plot_strings for s in validation_config.excluded_strings)
-
-
-@contextmanager
-def mock_settings_file(new_settings_file_path: Path, tmp_dir: Path):
-    shutil.copytree(main.views_helper.SETTINGS_PATH, tmp_dir, dirs_exist_ok=True)
-    shutil.copy(new_settings_file_path, tmp_dir)
-    with (
-        # Mocking is a bit more difficult because the values of default arguments are not overwritten once a function
-        # is imported, so it would not be enough just to overwrite SETTINGS_PATH
-        mock.patch.object(
-            main.views_helper.load_settings_from_file,
-            "__defaults__",
-            (
-                main.views_helper.load_settings_from_file.__defaults__[0],
-                tmp_dir.resolve(),
-            ),
-        ),
-        mock.patch.object(
-            ptm_vis_utils,
-            "CUSTOM_PTM_SETTINGS_FILE_STEM",
-            new_settings_file_path.stem,
-        ),
-    ):
-        yield
-
-
 class TestPTMVisualization:
     @pytest.fixture
     def evidence_df(self):
@@ -331,10 +155,52 @@ class TestPTMVisualization:
         return PlotValidationConfig()
 
     @pytest.fixture
-    def tau_casette_exon_config(self):
+    def tau_cassette_exon_config(self):
         return PlotValidationConfig(
-            required_ptm_types=("Phosphorylation",),
-            required_ptms=("S68", "T71", "S113"),
+            required_ptm_types=("Phosphorylation", "Ubiquitination", "Acetylation"),
+            required_ptms=(
+                "S68",
+                "T71",
+                "T111",
+                "S113",
+                "T175",
+                "T181",
+                "S185",
+                "S191",
+                "S199",
+                "S202",
+                "T205",
+                "S210",
+                "T212",
+                "S214",
+                "T217",
+                "T231",
+                "S235",
+                "S237",
+                "S238",
+                "K254",
+                "K257",
+                "S262",
+                "K267",
+                "K281",
+                "S289",
+                "K298",
+                "S305",  # TODO: shouldn't the mod also be on K305?
+                "K311",
+                "K317",
+                "K321",
+                "K353",
+                "T361",
+                "K369",
+                "Y394",
+                "S396",
+                "S400",
+                "T403",
+                "S404",
+                "S412",
+                "S416",
+                "S422",
+            ),
             required_region_names=(
                 "N-term",
                 "N1",
@@ -352,13 +218,64 @@ class TestPTMVisualization:
         )
 
     @pytest.fixture
-    def tau_substitution_config(self, tau_casette_exon_config):
+    def tau_substitution_config(self, tau_cassette_exon_config):
+        required_ptms = list(tau_cassette_exon_config.required_ptms)
+        required_ptms.pop(required_ptms.index("S68"))
+        required_ptms.pop(required_ptms.index("S305"))
+        required_ptms.append("K305")
+
         return PlotValidationConfig(
-            required_ptm_types=("Phosphorylation", "Ubiquitination", "Acetylation"),
-            required_ptms=("S113", "K305", "K311", "K317", "K321"),
-            required_region_names=tau_casette_exon_config.required_region_names,
+            required_ptm_types=tau_cassette_exon_config.required_ptm_types,
+            required_ptms=tuple(required_ptms),
+            required_region_names=tau_cassette_exon_config.required_region_names,
             required_region_short_names=(),
-            required_cleavages=("306",),
+            required_cleavages=("103", "105", "306"),
+        )
+
+    @pytest.fixture
+    def tau_5_config(self, tau_substitution_config):
+        required_ptms = list(tau_substitution_config.required_ptms)
+        last_ptm_index_before_cassette_exon = required_ptms.index("K267")
+        required_ptms = required_ptms[: last_ptm_index_before_cassette_exon + 1] + [
+            "K274",
+            "K280",
+            "K286",
+            "K290",
+            "K322",
+            "K338",
+            "S365",
+            "S369",
+            "S373",
+            "S381",
+            "S385",
+            "S391",
+            "T330",
+            "T372",
+            "Y363",
+        ]
+
+        region_names = list(tau_substitution_config.required_region_names)
+        region_names.pop(region_names.index("R3"))
+
+        return PlotValidationConfig(
+            required_ptm_types=tau_substitution_config.required_ptm_types,
+            required_ptms=tuple(required_ptms),
+            required_region_names=tuple(region_names),
+            required_region_short_names=(),
+            required_cleavages=("103", "105"),
+        )
+
+    @pytest.fixture
+    def tau_8_config(self, tau_cassette_exon_config):
+        required_ptms = list(tau_cassette_exon_config.required_ptms)
+        required_ptms.pop(required_ptms.index("S68"))
+
+        return PlotValidationConfig(
+            required_ptm_types=tau_cassette_exon_config.required_ptm_types,
+            required_ptms=tuple(required_ptms),
+            required_region_names=tau_cassette_exon_config.required_region_names,
+            required_region_short_names=(),
+            required_cleavages=("103", "105", "306"),
         )
 
     @pytest.fixture
@@ -373,20 +290,8 @@ class TestPTMVisualization:
 
     @staticmethod
     def test_plotting_functions(plot_func, kwargs, gfap_config):
-        result = plot_func(**kwargs)
-        assert len(result["plots"]) == 1
-        plot = result["plots"][0]
-
-        validate_plot_outputs(
-            plot,
-            plot_func,
-            all_groups=(
-                set(kwargs["metadata_df"]["Group"].unique())
-                if "metadata_df" in kwargs
-                else set()
-            ),
-            required_groups={"clean", "old", "exon"},
-            validation_config=gfap_config,
+        run_plot_and_validate(
+            plot_func, kwargs, gfap_config, {"clean", "old", "exon"}
         )
 
     @staticmethod
@@ -399,20 +304,8 @@ class TestPTMVisualization:
         }
         alter_ptm_settings(monkeypatch, new_param_dict)
 
-        result = plot_func(**kwargs)
-        assert len(result["plots"]) == 1
-
-        plot = result["plots"][0]
-        validate_plot_outputs(
-            plot,
-            plot_func,
-            all_groups=(
-                set(kwargs["metadata_df"]["Group"].unique())
-                if "metadata_df" in kwargs
-                else set()
-            ),
-            required_groups={"clean", "old", "exon"},
-            validation_config=gfap_config,
+        run_plot_and_validate(
+            plot_func, kwargs, gfap_config, {"clean", "old", "exon"}
         )
 
     @staticmethod
@@ -570,16 +463,9 @@ class TestPTMVisualization:
 
     @staticmethod
     def test_different_metadata_column(plot_func, bar_detail_kwargs, gfap_config):
-        metadata_df = bar_detail_kwargs["metadata_df"]
         bar_detail_kwargs["metadata_column"] = "Batch"
-        result = plot_func(**bar_detail_kwargs)
-        assert len(result["plots"]) == 1
-        validate_plot_outputs(
-            result["plots"][0],
-            create_bar_ptm_visualization,
-            all_groups=set(metadata_df[bar_detail_kwargs["metadata_column"]].unique()),
-            required_groups={"2", "3", "4"},
-            validation_config=gfap_config,
+        run_plot_and_validate(
+            plot_func, bar_detail_kwargs, gfap_config, {"2", "3", "4"}
         )
 
     @staticmethod
@@ -659,32 +545,19 @@ class TestPTMVisualization:
             )
 
     @staticmethod
-    def test_cassette_exon(plot_func, kwargs, tau_casette_exon_config):
+    def test_cassette_exon(plot_func, kwargs, tau_cassette_exon_config):
         kwargs["evidence_df"] = get_evidence_df(TAU_EVIDENCE_FILE_PATH)
         if "metadata_df" in kwargs:
             kwargs["metadata_df"] = get_metadata_df(TAU_METADATA_FILE_PATH)
         kwargs["fasta_file_path"] = Path(TAU_PATH / "uniprotkb_P10636_7_8.fasta")
         kwargs["regions_file_path"] = TAU_REGIONS_FILE_PATH
 
-        result = plot_func(**kwargs)
-        assert len(result["plots"]) == 1
-        plot = result["plots"][0]
-
-        all_groups = (
-            set(kwargs["metadata_df"]["Group"].unique())
-            if "metadata_df" in kwargs
-            else set()
-        )
-        validate_plot_outputs(
-            plot,
-            plot_func,
-            all_groups=all_groups,
-            required_groups={"AD"},
-            validation_config=tau_casette_exon_config,
+        run_plot_and_validate(
+            plot_func, kwargs, tau_cassette_exon_config, {"AD", "CTR"}
         )
 
     @staticmethod
-    def test_modification_at_first_location(
+    def test_modification_at_regions_starts_and_ends(
         plot_func, kwargs, tmp_ptm_settings_dir, gfap_config
     ):
         # TODO: bar plot orders ptms in alternative exon by number and doesn't take exon into account
@@ -743,8 +616,9 @@ class TestPTMVisualization:
             ignore_index=True,
         )
 
-        # First two pepitdes of Exon 2
+        # First two peptides of Exon 2
         mock_exon2_peptide = kwargs["evidence_df"].iloc[97]
+        mock_exon2_peptide["Protein ID"] = "P14136-3"
         mock_exon2_peptide["Sequence"] = "GGKST"
         mock_exon2_peptide["Modified sequence"] = "_G(ci)GKST_"
         mock_exon2_peptide["Modifications"] = "ci"
@@ -753,6 +627,7 @@ class TestPTMVisualization:
             ignore_index=True,
         )
         mock_exon2_peptide = kwargs["evidence_df"].iloc[97]
+        mock_exon2_peptide["Protein ID"] = "P14136-3"
         mock_exon2_peptide["Sequence"] = "GGKST"
         mock_exon2_peptide["Modified sequence"] = "_GG(ci)KST_"
         mock_exon2_peptide["Modifications"] = "ci"
@@ -789,6 +664,7 @@ class TestPTMVisualization:
 
         # End of sequence/exon 2
         mock_sequence_end_peptide = kwargs["evidence_df"].iloc[97]
+        mock_sequence_end_peptide["Protein ID"] = "P14136-3"
         mock_sequence_end_peptide["Sequence"] = "GTPPARG"
         mock_sequence_end_peptide["Modified sequence"] = "_GTPPARG(ci)_"
         mock_sequence_end_peptide["Modifications"] = "ci"
@@ -797,42 +673,30 @@ class TestPTMVisualization:
             ignore_index=True,
         )
 
+        additional_required_ptms = (
+            "M1",
+            "Q388",
+            "I389",
+            "R390",
+            "E391",
+            "G402",
+            "H403",
+            "M432",
+            "G391",
+            "G392",
+            "G431",
+        )
+        additional_excluded_strings = ("M0",)
+        gfap_config.required_ptms += additional_required_ptms
+        gfap_config.excluded_strings += additional_excluded_strings
+        gfap_config.required_region_short_names += ("α",)
+
         with mock_settings_file(
             TEST_PTM_VISUALIZATION_PATH / "ptm_settings_mods_at_first_location.yaml",
             tmp_ptm_settings_dir,
         ):
-            result = plot_func(**kwargs)
-            assert len(result["plots"]) == 1
-            plot = result["plots"][0]
-
-            additional_required_ptms = (
-                "M1",
-                "Q388",
-                "I389",
-                "R390",
-                "E391",
-                "G402",
-                "H403",
-                "M432",
-                "G391",
-                "G392",
-                "G431",
-            )
-            additional_excluded_strings = ("M0",)
-            gfap_config.required_ptms += additional_required_ptms
-            gfap_config.excluded_strings += additional_excluded_strings
-            gfap_config.required_region_short_names += ("α",)
-
-            validate_plot_outputs(
-                plot,
-                plot_func,
-                all_groups=(
-                    set(kwargs["metadata_df"]["Group"].unique())
-                    if "metadata_df" in kwargs
-                    else set()
-                ),
-                required_groups={"clean", "old", "exon"},
-                validation_config=gfap_config,
+            run_plot_and_validate(
+                plot_func, kwargs, gfap_config, {"clean", "old", "exon"}
             )
 
     @staticmethod
@@ -845,48 +709,38 @@ class TestPTMVisualization:
         if "metadata_df" in kwargs:
             kwargs["metadata_df"] = get_metadata_df(EXAMPLE_DATASET_METADATA_FILE)
 
-        result = plot_func(**kwargs)
-        assert len(result["plots"]) == 1
-        plot = result["plots"][0]
-
-        validate_plot_outputs(
-            plot,
-            plot_func,
-            all_groups=(
-                set(kwargs["metadata_df"]["Group"].unique())
-                if "metadata_df" in kwargs
-                else set()
-            ),
-            required_groups={"REL-FREE", "RELAPSE"},
-            validation_config=satb1_config,
+        run_plot_and_validate(
+            plot_func, kwargs, satb1_config, {"REL-FREE", "RELAPSE"}
         )
 
     @staticmethod
     def test_single_amino_acid_substitution_end_of_exon(
         plot_func, kwargs, tau_substitution_config
     ):
+        # TODO: region names and how they are separated are FUBAR
         kwargs["evidence_df"] = get_evidence_df(TAU_EVIDENCE_FILE_PATH)
         if "metadata_df" in kwargs:
             kwargs["metadata_df"] = get_metadata_df(TAU_METADATA_FILE_PATH)
         kwargs["fasta_file_path"] = TAU_PATH / "uniprotkb_P10636_5_8.fasta"
         kwargs["regions_file_path"] = TAU_REGIONS_FILE_PATH
 
-        result = plot_func(**kwargs)
-        assert len(result["plots"]) == 1
-        plot = result["plots"][0]
-
-        validate_plot_outputs(
-            plot,
-            plot_func,
-            all_groups=(
-                set(kwargs["metadata_df"]["Group"].unique())
-                if "metadata_df" in kwargs
-                else set()
-            ),
-            required_groups=(
-                {"AD", "CTR"}
-                if plot_func == create_details_ptm_visualization
-                else {"AD"}
-            ),
-            validation_config=tau_substitution_config,
+        run_plot_and_validate(
+            plot_func, kwargs, tau_substitution_config, {"AD", "CTR"}
         )
+
+    @staticmethod
+    def test_fasta_with_single_sequence(plot_func, kwargs, tau_8_config, tau_5_config):
+        # Set up common test data
+        kwargs["evidence_df"] = get_evidence_df(TAU_EVIDENCE_FILE_PATH)
+        if "metadata_df" in kwargs:
+            kwargs["metadata_df"] = get_metadata_df(TAU_METADATA_FILE_PATH)
+
+        # Tau-8
+        kwargs["fasta_file_path"] = TAU_PATH / "uniprotkb_P10636_8.fasta"
+        kwargs["regions_file_path"] = TAU_REGIONS_FILE_PATH
+        run_plot_and_validate(plot_func, kwargs, tau_8_config, {"AD", "CTR"})
+
+        # Tau-5
+        kwargs["fasta_file_path"] = TAU_PATH / "uniprotkb_P10636_5.fasta"
+        kwargs["regions_file_path"] = TAU_PATH / "regions_P10636_5.csv"
+        run_plot_and_validate(plot_func, kwargs, tau_5_config, {"AD", "CTR"})
